Switch from xAPIC (MMIO) to x2APIC (MSR-based) LAPIC access
Eliminates MMIO mapping, Lock<Option<Mmio>>, and LAPIC_BASE atomic.
All register access now via rdmsr/wrmsr. ICR writes are a single 64-bit
MSR write (race-free). APIC IDs widened to u32 with MADT type 9 parsing.

+++ kernel/src/arch/smp.rs
@@ -193,7 +193,7 @@ pub fn boot_aps(madt: &MadtInfo, boot_cr3: u64) {
 
         let ap_cpu_id = next_cpu_id;
         next_cpu_id += 1;
-        let ap_percpu = percpu::alloc_ap(ap_cpu_id, ap_id as u32);
+        let ap_percpu = percpu::alloc_ap(ap_cpu_id, ap_id);
 
         data.stack_top = stack_base as u64 + AP_STACK_SIZE as u64;
         data.entry = ap_entry as *const () as u64;
